Create Step2_MLPE_scaled.R
This code takes transformed HSM and calculates MLPE for each of the different connectivity metrics. The output is an excel (cvs) file that reports AIC, BIC, residual and marginal R2 values. This output is used to select the best performing transformation and metric (i.e., which combinations of HSM and connectivity best predict gene flow across the landscape)